Update README.md to use markdown syntax and simply procedure
Previously, README.md was actually a plain-text file that did not conform to markdown syntax. Turn the list into an actual list, and make the instructions more explicit by calling out commands.

Additionally, replace the "finishing" step shell script with a few explicit commands to create an working copy that is under the expected name.